replace feeze with backwards compatible freezeTag file
